feat: api 분리 및 명 변경, 페이징 처리 로직 수정
