+++ backend/src/main/java/com/backend/domain/chat/controller/ApiV1ChatController.java
@@ -2,24 +2,15 @@ package com.backend.domain.chat.controller;
 
 import static org.springframework.http.MediaType.*;
 
-import java.util.Map;
-
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.web.PageableDefault;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
-import org.springframework.messaging.handler.annotation.DestinationVariable;
-import org.springframework.messaging.handler.annotation.Header;
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.handler.annotation.Payload;
-import org.springframework.messaging.handler.annotation.SendTo;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
-import com.backend.domain.chat.dto.request.ChatRequest;
-import com.backend.domain.chat.dto.response.ChatResponse;
 import com.backend.domain.chat.dto.response.ChatResponses;
 import com.backend.domain.chat.dto.response.ChatsInPost;
 import com.backend.domain.chat.service.ChatService;
@@ -29,19 +20,13 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-@Controller
+@RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/chat")
-public class ChatV1Controller {
+public class ApiV1ChatController {
 
 	private final ChatService chatService;
 
-	@GetMapping()
-	public String home() {
-		return "index";
-	}
-
-	@ResponseBody
 	@GetMapping(value = "/list/{postId}", produces = APPLICATION_JSON_VALUE)
 	public GenericResponse<ChatResponses> getChattingList(
 		@PathVariable(name = "postId") Long postId) {
@@ -50,24 +35,17 @@ public class ChatV1Controller {
 		return GenericResponse.of(true, HttpStatus.OK.value(), result, "요청 성공");
 	}
 
-	@ResponseBody
 	@GetMapping(value = "/page/{postId}", produces = APPLICATION_JSON_VALUE)
 	public GenericResponse<ChatsInPost> getChattingList(
 		@PathVariable(name = "postId") Long postId,
-		@PageableDefault(sort = "createdAt") Pageable pageable) {
+		@RequestParam(value = "page", defaultValue = "0") int pageNum,
+		@RequestParam(value = "size", defaultValue = "10") int pageSize) {
 
-		ChatsInPost result = chatService.getByPostId(postId, pageable);
+		ChatsInPost result = chatService.getByPostId(postId, PageRequest.of(pageNum, pageSize,
+			Sort.by(Sort.Direction.ASC, "createdAt"))
+		);
 
 		return GenericResponse.of(true, HttpStatus.OK.value(), result, "요청 성공");
-	}
-
-	@MessageMapping("/msg/{postId}")
-	@SendTo("/topic/{postId}")
-	public ChatResponse sendMessage(@DestinationVariable Long postId,
-		@Header("simpSessionAttributes") Map<String, Object> sessionAttributes,
-		@Payload ChatRequest chatRequest) {
-
-		return chatService.save(chatRequest, postId, sessionAttributes);
 	}
 
 }
