+++ backend/src/test/java/com/backend/domain/chat/ChatControllerTest.java
@@ -24,7 +24,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 @Sql(scripts = {"/sql/init.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_CLASS)
 @Sql(scripts = {"/sql/delete.sql"}, executionPhase = Sql.ExecutionPhase.AFTER_TEST_CLASS)
-class ChatV1ControllerTest {
+class ApiV1ChatControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
